+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - A region's reports are now accessible for the ambassadors in charge via the region menus !1041 @peter.toennies
 - Add basket counters to statistics (issue #81) !1045 @chris2up9
 - Add basket counters to statistics !1045 @chris2up9
+- Placed event buttons correctly for mobile on dashboard #640 !1044 @henrikhertler
 
 ## Bugfixes
 - Profile button "remove from all slots" is now only enabled for orga !968 #362 @fs_k